+++ src/collector/dnd.js
@@ -1,7 +1,6 @@
-import $ from 'jquery';
-import Emitter from '../emitter';
-import Runtime from '../html5/runtime';
-import File from '../file';
+const Emitter = require('../emitter');
+const Runtime = require('../html5/runtime');
+const File = require('../file');
 
 const MAX_NUM_ONCE = 100;
 
@@ -85,12 +84,11 @@ class Area extends Emitter {
 const Collectors = [];
 
 function prepare() {
-    if (!('DataTransfer' in window) || !('FileList' in window)) {
+    if (!('DataTransfer' in window) || !('FileList' in window) || !document.addEventListener) {
         return;
     }
 
-    const $doc = $(document),
-        runtime = Runtime.getInstance();
+    const runtime = Runtime.getInstance();
 
     let started = 0, enter = 0, endTimer;
     const dataTransferReader = createReader((file, responders) => {
@@ -119,7 +117,7 @@ function prepare() {
     const move = (e) => {
         const has = Collectors.filter(responder => responder.response(e)).length > 0;
 
-        const dataTransfer = (e.originalEvent || e).dataTransfer;
+        const dataTransfer = e.dataTransfer;
 
         if (dataTransfer) {
             dataTransfer.dropEffect = has ? 'copy' : 'none';
@@ -156,7 +154,7 @@ function prepare() {
             return;
         }
 
-        let dataTransfer = (e.originalEvent || e).dataTransfer;
+        let dataTransfer = e.dataTransfer;
 
         try {
             if (dataTransfer.getData('text/html')) {
@@ -167,11 +165,13 @@ function prepare() {
         dataTransferReader(dataTransfer, responders);
     };
 
-    $doc.on('dragenter dragover dragleave', drag);
-    $doc.on('drop', drop);
+    document.addEventListener('dragenter', drag, false);
+    document.addEventListener('dragover', drag, false);
+    document.addEventListener('dragleave', drag, false);
+    document.addEventListener('drop', drop, false);
 }
 
-export default class DndCollector {
+class DndCollector {
 
     constructor(core) {
         if (Collectors.length < 1) {
@@ -222,3 +222,5 @@ export default class DndCollector {
         this.areas.forEach((area) => area.end(e));
     }
 }
+
+module.exports = DndCollector;