+++ src/collector/picker.js
@@ -1,27 +1,26 @@
-import $ from 'jquery';
-import Emitter from '../emitter';
-import Html5Runtime from '../html5/runtime';
-import FlashRuntime from '../flash/runtime';
-import File from '../file';
+const Emitter = require('../emitter');
+const Html5Runtime = require('../html5/runtime');
+const FlashRuntime = require('../flash/runtime');
+const File = require('../file');
+const {extend} = require('../util');
+
 
 let SWF_URL = '';
 
+const createElement = (() => {
+    const div = document.createElement('div');
+    return (html) => {
+        div.innerHTML = html;
+        html = div.firstChild;
+        div.removeChild(html);
+        return html;
+    };
+})();
+
 class FlashTriggerCollection {
     constructor(core, onFiles) {
-        let overlay = $('<label></label>');
-        overlay.css({
-            position: 'fixed',
-            opacity: 0,
-            left: -100,
-            top: -100,
-            width: 50,
-            height: 50,
-            display: 'block',
-            cursor: 'pointer',
-            background: 'black',
-            overflow: 'hidden',
-            zIndex: 99999
-        });
+        const overlay = createElement('<label style="position:fixed;left:-100px;top:-100px;width:50px;height:50px;display:block;cursor:pointer;overflow:hidden;z-index:99999;opacity:0;filter:alpha(opacity=0)"></label>');
+
         const runtime = new FlashRuntime(overlay, SWF_URL, () => {
             return {
                 accept: core.getAccept(),
@@ -35,15 +34,17 @@ class FlashTriggerCollection {
         });
 
         runtime.on('rollOut', () => this.hideOverlay());
-        this.overlay = overlay.appendTo(document.body);
+
+        document.body.appendChild(overlay);
+        this.overlay = overlay;
     }
 
     hideOverlay() {
-        this.overlay.css({
-            left: -100,
-            top: -100,
-            width: 50,
-            height: 50
+        extend(this.overlay.style, {
+            left: '-100px',
+            top: '-100px',
+            width: '50px',
+            height: '50px'
         });
         if (this.current) {
             this.current.emit('rollOut');
@@ -54,95 +55,110 @@ class FlashTriggerCollection {
 
     add(trigger) {
         const overlay = this.overlay, emitter = new Emitter;
-        trigger = trigger.on ? trigger : $(trigger);
-        trigger.on('mouseover.flashpicker', (e) => {
-            let rect = e.currentTarget.getBoundingClientRect();
-            overlay.css({
-                left: rect.left,
-                top: rect.top,
-                width: rect.right - rect.left,
-                height: rect.bottom - rect.top
+        const mouseover = () => {
+            const rect = trigger.getBoundingClientRect();
+            extend(overlay.style, {
+                left: rect.left + 'px',
+                top: rect.top + 'px',
+                width: (rect.right - rect.left) + 'px',
+                height: (rect.bottom - rect.top) + 'px'
             });
             emitter.emit('rollOver');
             if (this.current && this.current !== emitter) {
                 this.current.emit('rollOut');
             }
             this.current = emitter;
-        });
+        };
+
+        if (trigger.addEventListener) {
+            trigger.addEventListener('mouseover', mouseover, false);
+        } else if (trigger.attachEvent) {
+            trigger.attachEvent("onmouseover", mouseover);
+        }
 
         emitter.destroy = () => {
             if (this.current === emitter) {
                 this.hideOverlay();
             }
             emitter.removeAllListeners();
-            trigger.off('mouseover.flashpicker');
+            if (trigger.removeEventListener) {
+                trigger.removeEventListener('mouseover', mouseover, false);
+            } else if (trigger.detachEvent) {
+                trigger.detachEvent("onmouseover", mouseover);
+            }
         };
 
         return emitter;
+    }
+}
+
+class Html5Trigger extends Emitter {
+    constructor(trigger, core, onFiles) {
+        super();
+
+        this.trigger = trigger;
+        this.core = core;
+        this.label = createElement('<label style="position:absolute;top:0;left:0;width:100%;height:100%;display:inline-block;cursor:pointer;background:#fff;overflow:hidden;opacity:0"></label>');
+
+        this.onChange = (e) => {
+            onFiles(e.target.files);
+            this.destroyInput();
+            this.createInput();
+        };
+        trigger.appendChild(this.label);
+        this.createInput();
+    }
+
+    createInput() {
+        const input = createElement('<input type="file" style="position:absolute;clip:rect(1px 1px 1px 1px);" />');
+
+        let accept = this.core.getAccept();
+        if (accept && accept.length > 0) {
+            accept = accept.map((item) => {
+                return item.mimeTypes || ('.' + item.extensions.join(',.'));
+            });
+
+            input.setAttribute('accept', accept.join(','));
+        }
+        if (this.core.isMultiple()) {
+            input.setAttribute('multiple', 'multiple');
+        }
+
+        input.addEventListener('change', this.onChange, false);
+
+        this.label.appendChild(input);
+        this.input = input;
+    }
+
+    destroyInput() {
+        if (!this.input) return;
+        this.input.removeEventListener('change', this.onChange, false);
+        this.label.removeChild(this.input);
+        this.input = null;
+    }
+
+    destroy() {
+        this.destroyInput();
+        this.removeAllListeners();
+        this.trigger.removeChild(this.label);
     }
 }
 
 class Html5TriggerCollection {
     constructor(core, onFiles) {
         const runtime = Html5Runtime.getInstance();
-
-        this._createInput = (label) => {
-            let input = $(document.createElement('input'));
-
-            input.attr('type', 'file');
-            input.css({
-                position:'absolute',
-                clip: 'rect(1px 1px 1px 1px)'
-            });
-
-            let accept = core.getAccept();
-            if (accept && accept.length > 0) {
-                accept = accept.map((item) => {
-                    return item.mimeTypes || ('.' + item.extensions.join(',.'));
-                });
-
-                input.attr('accept', accept.join(','));
-            }
-            if (core.isMultiple()) {
-                input.attr('multiple', 'multiple');
-            }
-
-            input.on('change', (e) => {
-                onFiles(e.target.files, runtime);
-                input.remove();
-                this._createInput(label);
-            });
-            label.html(input);
+        this.core = core;
+        this.onFiles = (files) => {
+            onFiles(files, runtime);
         };
     }
 
     add(trigger) {
-        const label = $('<label></label>'), emitter = new Emitter;
-        label.css({
-            position: 'absolute',
-            opacity: 0,
-            top:0,left:0,
-            width: '100%',
-            height: '100%',
-            display: 'inline-block',
-            cursor: 'pointer',
-            background: '#fff',
-            overflow: 'hidden'
-        });
-        this._createInput(label);
-        trigger = trigger.append ? trigger : $(trigger);
-        trigger.append(label);
-
-        emitter.destroy = () => {
-            emitter.removeAllListeners();
-            label.remove();
-        };
-
-        return emitter;
+        return new Html5Trigger(trigger, this.core, this.onFiles);
     }
 }
 
-export default class PickerCollector {
+class PickerCollector {
     static setSWF(url) {
         SWF_URL = url;
     }
@@ -167,3 +183,5 @@ export default class PickerCollector {
         return this.triggerCollection.add(area);
     }
 }
+
+module.exports = PickerCollector;